Rename METS scripts appropriately
