+++ src/MCPClient/lib/clientScripts/v1_create_mets/create_mets_trim.py
