+++ src/MCPClient/lib/clientScripts/v1_create_mets/v1_create_aip_mets.py
@@ -31,16 +31,16 @@ from main.models import (
     SIPArrange,
 )
 
-import archivematicaCreateMETSReingest
-from archivematicaCreateMETSMetadataCSV import parseMetadata
-from archivematicaCreateMETSRights import archivematicaGetRights
-from archivematicaCreateMETSRightsDspaceMDRef import (
+import create_mets_reingest
+from create_mets_metadata_csv import parseMetadata
+from create_mets_rights import archivematicaGetRights
+from create_mets_rights_dspace_mdref import (
     archivematicaCreateMETSRightsDspaceMDRef,
 )
-from archivematicaCreateMETSTrim import getTrimDmdSec
-from archivematicaCreateMETSTrim import getTrimFileDmdSec
-from archivematicaCreateMETSTrim import getTrimAmdSec
-from archivematicaCreateMETSTrim import getTrimFileAmdSec
+from create_mets_trim import getTrimDmdSec
+from create_mets_trim import getTrimFileDmdSec
+from create_mets_trim import getTrimAmdSec
+from create_mets_trim import getTrimFileAmdSec
 
 # archivematicaCommon
 from archivematicaFunctions import escape
@@ -1614,7 +1614,7 @@ def create_mets(job, opts):
         job.pyprint("Updating METS during reingest")
         # fileGroupIdentifier is SIPUUID, baseDirectoryPath is SIP dir,
         # don't keep existing normative structmap if creating one
-        root = archivematicaCreateMETSReingest.update_mets(
+        root = create_mets_reingest.update_mets(
             job,
             baseDirectoryPath,
             fileGroupIdentifier,
